feature/54 refactor: 패키지 분리 및 geolocation timeout 변경
