+++ front/src/components/post/CommentInput.vue
@@ -15,7 +15,7 @@
 </template>
 
 <script>
-import router from "../router";
+import router from "../../router";
 
 export default {
   name: "CommentInput",
