+++ front/src/components/post/PostDetailPage.vue
